requirements.txt
numpy
pandas
python-dotenv
mysql-connector-python
pymysql
scikit-learn
seaborn
matplotlib
warnings
catboost
xgboost
dataclasses
pickle
dill
mlflow==2.2.2





# -e .